requirements.txt
click==8.0.4
colorama==0.4.4
Flask==2.0.3
Flask-Login==0.5.0
Flask-MySQLdb==1.0.1
Flask-WTF==1.0.0
greenlet==1.1.2
gunicorn==20.1.0
itsdangerous==2.1.0
Jinja2==3.0.3
MarkupSafe==2.1.0
mysqlclient==2.1.0
numpy==1.22.2
pandas==1.4.1
PyMySQL==1.0.2
python-dateutil==2.8.2
pytz==2021.3
scipy==1.8.0
six==1.16.0
SQLAlchemy==1.4.31
Werkzeug==2.0.3
WTForms==3.0.1
